+++ packages/gamut/src/GridForm/GridFormInputGroup/GridFormTextInput/__tests__/GridFormTextInput-test.tsx
@@ -1,21 +1,47 @@
+import { setupRtl } from '@codecademy/gamut-tests';
+import userEvent from '@testing-library/user-event';
+
+import { stubTextField } from '../../../__tests__/stubs';
 import { itHandlesAriaInvalid } from '../../__fixtures__/assertions';
-import { renderGridFormTextInput } from '../../__fixtures__/renderers';
+import { GridFormTextInput } from '..';
+
+const renderGridFormTextInput = setupRtl(GridFormTextInput, {
+  clearErrors: jest.fn(),
+  field: stubTextField,
+  register: jest.fn(),
+});
 
 describe('GridFormTextInput', () => {
   describe('when an id is passed as a prop', () => {
     it('renders an input with the same id', () => {
-      const textInput = renderGridFormTextInput({ id: 'mycoolid' });
+      const { view } = renderGridFormTextInput({
+        field: {
+          ...stubTextField,
+          id: 'mycoolid',
+        },
+      });
 
-      expect(textInput.find('input#mycoolid').length).toBe(1);
+      expect(view.getByRole('textbox')).toHaveProperty('id', 'mycoolid');
     });
   });
 
   describe('when no id is passed', () => {
     it('renders an input with the id equal to the field name', () => {
-      const textInput = renderGridFormTextInput({ name: 'name' });
+      const { view } = renderGridFormTextInput({
+        field: {
+          ...stubTextField,
+          name: 'name',
+        },
+      });
 
-      expect(textInput.find('input#name').length).toBe(1);
+      expect(view.getByRole('textbox')).toHaveProperty('id', 'name');
     });
+  });
+
+  it('clears errors when typed into', () => {
+    const { view } = renderGridFormTextInput();
+
+    userEvent.type(view.getByRole('textbox'), 'hi');
   });
 
   itHandlesAriaInvalid('GridFormTextInput', 'input');